fix: Fix timestamp management in unit tests
If the event has a `@Timestamp` annotation, Drools will use that field to determine when the event had occurred. It's a "read-only" type of usage because if we set the pseudo-clock like we do in the tests, that won't affect the actual class. Drools won't update the timestamp upon inserting for us. That only happens if the class doesn't have a `@Timestamp`. But we currently need the events to have a `Date timestamp` that should be "sync'd" with Drools.
The Drools documentation isn't very clear about this and this was a pain to debug. Maybe in the future our events won't need a `Date timestamp`.