feat: Write FailedLogin event and block for FailedLogin
The new events don't have a `@Timestamp` annotation nor are Spring JPA Entities. I'm wondering if this is the way to go. No point in keeping the same thing in both DB and Drools unless we explicitly need it.
The BlockEvent is 24h only. Apparently there's no way to parameterize this.