feat: Give block events a duration
By taking advantage of the `@Duration` annotation, block events can have arbitrary length. Unfortunately, I had to remove `@Expires` becuase that's still hard-coded _unless_ we have a pre-defined maximum block duration.

Also, when an event expires, it's removed from the memory and `not BlockEvent(...)` works as it should. On the other hand, if an event finished (i.e. `@Duration` ends), this is not the case so `not BlockEvent` isn't enough: we have to check that a BlockEvent doesn't exist OR it does exist but it's `before SomethingElse`, that `SomethingElse` being a new event that was inserted into the session. AFAIK There is no other way to do this.