feat: Write data init method for the server (instead of `data.sql`)
I want to initialize roles, permissions and users. We can't use data.sql because that's SQL only - and here we also want to write all the facts in the rule engine.

I also removed the ID field in Role and Permission ctors because otherwise Spring JPA won't autgenerate them upon saving to the DB: it has to be null.