feat: Nag user if his password is weak
Much like in smartcityhub, we keep the nagPassword field in the JWT.
Upon a successful login, we check if the plaintext password used is in the working memory of WeakPassword-s, and if so, update the JWT accordingly.
root.tsx contains the lifecycle hook that pops a toast message if the password is too weak.

I lowered the criteria needed to trigger a WeakPassword event (2 accounts with same password) so that testing is easier. Launch the make_weak_password script to make `john123` weak.